chore: pass workExamples through job configuration
This removes an unwanted dependency between the cli and api packages.

It has the side-effect of providing the work examples before the job is
finished. Not sure if it's an issue.
