+++ apis/core/lib/domain/job/create.ts
@@ -1,13 +1,17 @@
+import $rdf from 'rdf-ext'
 import { cc } from '@cube-creator/core/namespace'
 import { GraphPointer } from 'clownface'
 import { NamedNode, Term } from 'rdf-js'
+import { DESCRIBE } from '@tpluscode/sparql-builder'
 import * as Job from '@cube-creator/model/Job'
 import * as ImportJob from '@cube-creator/model/ImportJob'
-import { CsvMapping, Project, Dataset, ImportProject } from '@cube-creator/model'
+import { CsvMapping, Project, Dataset, ImportProject, CsvProject } from '@cube-creator/model'
 import { ResourceStore } from '../../ResourceStore'
 import * as id from '../identifiers'
 import { DomainError } from '@cube-creator/api-errors'
+import env from '@cube-creator/core/env'
 import { schema } from '@tpluscode/rdf-ns-builders'
+import { isCsvProject } from '@cube-creator/model/Project'
 
 interface StartTransformationCommand {
   resource: NamedNode
@@ -55,7 +59,7 @@ export async function createPublishJob({
     throw new Error('Project not found from jobs collection')
   }
 
-  const project = await store.getResource<Project>(projectPointer)
+  const project = await store.getResource<CsvProject | ImportProject>(projectPointer)
   const organization = await store.getResource(project.maintainer)
 
   if (!organization.publishGraph) {
@@ -65,7 +69,7 @@ export async function createPublishJob({
   const metadata = await store.getResource(project.dataset)
 
   const jobPointer = await store.createMember(jobCollection.term, id.job(jobCollection))
-  Job.createPublish(jobPointer, {
+  const job = Job.createPublish(jobPointer, {
     project: projectPointer,
     name: 'Publish job',
     revision: project.nextRevision,
@@ -74,7 +78,73 @@ export async function createPublishJob({
     publishedTo: metadata?.pointer.out(schema.workExample).term,
   })
 
+  const cubeId = isCsvProject(project)
+    ? organization.createIdentifier({
+      cubeIdentifier: project.cubeIdentifier,
+    })
+    : project.sourceCube
+
+  const dataset = await store.getResource<Dataset>(project.dataset.id)
+  const workExamples = prepareWorkExamples(cubeId, job, dataset)
+  workExamples.forEach(workExample => job.addWorkExample(workExample))
+
   return jobPointer
+}
+
+export function prepareWorkExamples(cubeId: NamedNode, job: Job.PublishJob, dataset: Dataset): Job.WorkExampleInput[] {
+  const workExamples = []
+
+  if (env.has('TRIFID_UI')) {
+    workExamples.push({
+      url: $rdf.namedNode(lindasWebQueryLink(cubeId, job.revision, job.publishGraph)),
+      encodingFormat: $rdf.literal('Application/Sparql-query'),
+      name: [
+        $rdf.literal('SPARQL Endpoint mit Vorauswahl des Graph', 'de'),
+        $rdf.literal('SPARQL Endpoint with graph preselection', 'en'),
+        $rdf.literal('SPARQL Endpoint avec présélection du graphe', 'fr'),
+      ],
+    })
+  }
+
+  const visualize = $rdf.namedNode('https://ld.admin.ch/application/visualize')
+  const isPublishedToVisualize = dataset.pointer.has(schema.workExample, visualize).terms.length > 0
+  if (env.has('VISUALIZE_UI') && isPublishedToVisualize) {
+    workExamples.push({
+      url: $rdf.namedNode(visualizeLink(cubeId)),
+      encodingFormat: $rdf.literal('text/html'),
+      name: [
+        $rdf.literal('visualize.admin.ch'),
+      ],
+    })
+  }
+
+  return workExamples
+}
+
+function lindasWebQueryLink(cube: NamedNode, version: number, cubeGraph: NamedNode) {
+  const describe = DESCRIBE`?cube`
+    .FROM(cubeGraph)
+    .WHERE`
+      ${cube} ${schema.hasPart} ?cube .
+      ?cube ${schema.version} ${version} .
+    `.build()
+
+  const queryLink = new URL(env.TRIFID_UI)
+  queryLink.hash = new URLSearchParams([
+    ['query', describe],
+    ['endpoint', env.PUBLIC_QUERY_ENDPOINT],
+  ]).toString()
+
+  return queryLink.toString()
+}
+
+function visualizeLink(cube: NamedNode) {
+  const language = 'en'
+  const visualizeLink = new URL(`/${language}/create/new`, env.VISUALIZE_UI)
+
+  visualizeLink.searchParams.set('cube', cube.value)
+
+  return visualizeLink.toString()
 }
 
 export async function createImportJob({ store, resource }: StartImportCommand): Promise<GraphPointer<NamedNode>> {
