+++ cli/lib/metadata.ts
@@ -1,5 +1,5 @@
 import { obj } from 'through2'
-import { Literal, NamedNode, Quad, Quad_Object as QuadObject, Quad_Subject as QuadSubject, Term } from 'rdf-js'
+import { Literal, NamedNode, Quad, Quad_Object as QuadObject, Quad_Subject as QuadSubject } from 'rdf-js'
 import $rdf from 'rdf-ext'
 import { dcterms, rdf, schema, sh, xsd } from '@tpluscode/rdf-ns-builders'
 import { cc, cube } from '@cube-creator/core/namespace'
@@ -10,7 +10,6 @@ import type { Context } from 'barnard59-core/lib/Pipeline'
 import { loadDimensionMapping } from './output-mapper'
 import TermMap from '@rdfjs/term-map'
 import { Published } from '@cube-creator/model/Cube'
-import { prepareWorkExamples } from '../../apis/core/lib/domain/job/update'
 
 export async function loadDataset(jobUri: string, Hydra: HydraClient) {
   const jobResource = await Hydra.loadResource<PublishJob>(jobUri)
@@ -85,17 +84,16 @@ export async function injectMetadata(this: Context, jobUri: string) {
         this.push($rdf.quad(maintainer.dataset, schema.dataset, quad.subject))
       }
 
-      // Links to visualize and lindas
-      const workExamples = prepareWorkExamples(baseCube, job, dataset)
-      workExamples.forEach(workExampleData => {
+      // Copy workExamples from job
+      const predicatesToCopy = [rdf.type, schema.name, schema.url, schema.encodingFormat]
+      job.workExamples.forEach(jobWorkExample => {
         const workExample = $rdf.blankNode()
         this.push($rdf.quad(quad.subject, schema.workExample, workExample))
-        this.push($rdf.quad(workExample, rdf.type, schema.CreativeWork))
-        workExampleData.name.forEach(name => {
-          this.push($rdf.quad(workExample, schema.name, name))
+        predicatesToCopy.forEach(predicate => {
+          jobWorkExample.pointer.out(predicate).forEach(object => {
+            this.push($rdf.quad(workExample, predicate, object.term as NamedNode | Literal))
+          })
         })
-        this.push($rdf.quad(workExample, schema.url, workExampleData.url))
-        this.push($rdf.quad(workExample, schema.encodingFormat, workExampleData.encodingFormat))
       });
 
       [...datasetTriples.match(dataset.id)]
